requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
faiss-cpu
sentence-transformers
unstructured
